// projects/openapi-e2e-test/src/generated/partial-content/requestServerTypes.ts
/*
 * This file has been generated by Oats, please don't modify it by hand!
 *
 * Generated from schemas/partial-content.json (originating from oats-ts/oats-schemas)
 */

import { Try } from '@oats-ts/openapi-runtime'

export type OptionalRequestBodyServerRequest = {
  mimeType?: 'application/json'
  body: Try<
    | {
        foo?: string
      }
    | undefined
  >
}