// projects/openapi-e2e-test/src/generated/partial-content/responseServerTypes.ts
/*
 * This file has been generated by Oats, please don't modify it by hand!
 *
 * Generated from schemas/partial-content.json (originating from oats-ts/oats-schemas)
 */

export type MissingBodyServerResponse = {
  statusCode: 200
}

export type OptionalRequestBodyServerResponse = {
  statusCode: 200
  mimeType: 'application/json'
  body: {
    foo?: string
  }
}